pyproject.toml
[project]
name = "wukong-robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8.10"
dependencies = [
    "apscheduler>=3.11.2",
    "asyncio>=4.0.0",
    "baidu-aip==2.0.0.1",
    "cryptography>=46.0.3",
    "edge-tts>=7.2.7",
    "faiss-cpu>=1.8.0.post1",
    "fire==0.1.3",
    "funasr-onnx>=0.4.1",
    "jieba>=0.42.1",
    "jsonpath>=0.82.2",
    "markdown==3.0.1",
    "nest-asyncio>=1.6.0",
    "openai>=2.2.0",
    "paho-mqtt>=2.1.0",
    "pip>=25.0.1",
    "pvporcupine>=3.0.4",
    "pvrecorder==1.1.1",
    "pyaudio>=0.2.14",
    "pydub==0.23.1",
    "pypinyin>=0.55.0",
    "python-dateutil==2.7.5",
    "pytz==2018.9",
    "pyyaml>=4.2b1",
    "requests==2.31.0",
    "semver==2.8.1",
    "speechbrain>=1.0.3",
    "tencentcloud-sdk-python>=3.1.21",
    "torch<2.5.0",
    "torchaudio<2.5.0",
    "tornado==6.3.3",
    "watchdog==0.9.0",
    "websocket==0.2.1",
    "websocket-client>=1.8.0",
]
